+++ vortex-duckdb/src/convert/array/data_chunk_adaptor.rs
@@ -1,9 +1,26 @@
 use duckdb::core::{ArrayVector, DataChunkHandle, FlatVector, ListVector, StructVector};
 use duckdb::vtab::arrow::WritableVector;
+use vortex_dtype::FieldNames;
 
 pub struct DataChunkHandleSlice<'a> {
     chunk: &'a mut DataChunkHandle,
     column_index: usize,
+}
+
+/// A wrapper around a [`DataChunkHandle`] that extra info to create a vortex array
+pub struct NamedDataChunk<'a> {
+    pub chunk: &'a DataChunkHandle,
+    pub nullable: Option<&'a [bool]>,
+    pub names: Option<FieldNames>,
+}
+
+/// Since duckdb vectors only have a capacity, not a size this wrapper exists to allow the creation
+/// of a vortex array from a duckdb vector.
+/// Nullability is also included since the duckdb doesn't have this info its on the table.
+pub struct SizedFlatVector {
+    pub vector: FlatVector,
+    pub nullable: bool,
+    pub len: usize,
 }
 
 impl<'a> DataChunkHandleSlice<'a> {
@@ -30,5 +47,31 @@ impl WritableVector for DataChunkHandleSlice<'_> {
 
     fn list_vector(&mut self) -> ListVector {
         self.chunk.list_vector(self.column_index)
+    }
+}
+
+impl<'a> NamedDataChunk<'a> {
+    pub fn from_chunk(chunk: &'a DataChunkHandle) -> Self {
+        Self {
+            chunk,
+            nullable: None,
+            names: None,
+        }
+    }
+
+    pub fn named_chunk(chunk: &'a DataChunkHandle, names: FieldNames) -> Self {
+        Self {
+            chunk,
+            nullable: None,
+            names: Some(names),
+        }
+    }
+
+    pub fn new(chunk: &'a DataChunkHandle, nullable: &'a [bool], names: FieldNames) -> Self {
+        Self {
+            chunk,
+            nullable: Some(nullable),
+            names: Some(names),
+        }
     }
 }
